retoque finales para la entrega

+++ src/recipes/recipes.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, HttpException, ParseUUIDPipe, HttpStatus } from '@nestjs/common';
 import { RecipesService } from './recipes.service';
 import { CreateRecipeDto } from './dto/create-recipe.dto';
 import { UpdateRecipeDto } from './dto/update-recipe.dto';
@@ -11,33 +11,69 @@ export class RecipesController {
   constructor(private readonly recipesService: RecipesService) {}
 
   @Get('recommended')
-  GetRecipesByIngredients(@Body() IngredientsRecipeDto:IngredientsRecipeDto) {
-    return this.recipesService.getRecipesByIngredients(IngredientsRecipeDto)
+  async GetRecipesByIngredients(@Body() IngredientsRecipeDto:IngredientsRecipeDto) {
+    const response = await this.recipesService.getRecipesByIngredients(IngredientsRecipeDto);
+    if (response.length === 0) {
+      throw new HttpException('RECOMMENDED_NOT_FOUND', 404)
+    } else {
+      return response
+    } 
   }
 
   @Post()
-  create(@Body() createRecipeDto: CreateRecipeDto) {
-    return this.recipesService.create(createRecipeDto);
+  async create(@Body() createRecipeDto: CreateRecipeDto) {
+    const response = await this.recipesService.create(createRecipeDto);
+    if (response === false) {
+      throw new HttpException('BAD_REQUEST', 400)
+    } else {
+      return response
+    } 
   }
 
   @Get()
-  findAll() {
-    return this.recipesService.findAll();
+  async findAll() {
+    const response = await this.recipesService.findAll();
+    if (response.length === 0) {
+      throw new HttpException('RECIPES_NOT_FOUND', 404)
+    } else {
+      return response
+    } 
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.recipesService.findOne(id);
+  async findOne(@Param('id', ParseUUIDPipe) id: string) {
+    const response = await this.recipesService.findOne(id);
+    if (response.length === 0) {
+      throw new HttpException('RECIPE_NOT_FOUND', 404)
+    } else {
+      return response
+    } 
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateRecipeDto: UpdateRecipeDto) {
-    return this.recipesService.update(id, updateRecipeDto);
+  async update(@Param('id',ParseUUIDPipe) id: string, @Body() updateRecipeDto: UpdateRecipeDto) { 
+    const confirmedOfExistence = await this.recipesService.findOne(id);
+    if (confirmedOfExistence.length !== 0) {
+      const response = await this.recipesService.update(id, updateRecipeDto);
+      if (response === false) {
+        throw new HttpException('BAD_REQUEST', 400)
+      } else {
+        return response
+      } 
+    } else throw new HttpException('RECIPE_NOT_FOUND', 404)
   }
 
-  @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.recipesService.remove(id);
+  @Delete(':id',)
+  async remove(@Param('id',new ParseUUIDPipe({errorHttpStatusCode: HttpStatus.NOT_ACCEPTABLE})) id: string) {
+    const confirmedOfExistence = await this.recipesService.findOne(id);
+    if (confirmedOfExistence.length !== 0) {
+      const response = await this.recipesService.remove(id);
+    if (response === false) {
+      throw new HttpException('BAD_REQUEST', 400)
+    } else {
+      return response
+    } 
+  } else throw new HttpException('RECIPE_NOT_FOUND', 404)
   }
 
 }
